Extract plot hook repository slice
