+++ changelog.d/+plot-hook-repository-slice.changed.md
@@ -1,0 +1,2 @@
+Moved character plot hook and plot hook interest persistence into a dedicated
+repository slice while preserving the forum repository facade.
